BE não permite apagar, reativar ou suspender o Admin

+++ backend/src/main/java/aor/paj/dao/UserDao.java
@@ -24,7 +24,7 @@ public class UserDao {
 
     public boolean delete(Long id) {
         UserEntity user = findById(id);
-        if (user != null) {
+        if (user != null && !user.isAdmin()) {
             entityManager.remove(user);
             return true;
         }
@@ -33,7 +33,7 @@ public class UserDao {
 
     public boolean suspendUser(Long id) {
         UserEntity user = findById(id);
-        if (user != null && user.isActive()) {
+        if (user != null && user.isActive() && !user.isAdmin()) {
             user.setActive(false);
             entityManager.merge(user);
             return true;
@@ -43,7 +43,7 @@ public class UserDao {
 
     public boolean activateUser(Long id) {
         UserEntity user = findById(id);
-        if (user != null && !user.isActive()) {
+        if (user != null && !user.isActive() && !user.isAdmin()) {
             user.setActive(true);
             entityManager.merge(user);
             return true;
